+++ src/Model/AnonymousIp.php
@@ -31,45 +31,14 @@ use GeoIp2\Util;
  */
 class AnonymousIp extends AbstractModel
 {
-    /**
-     * @var bool
-     */
-    protected $isAnonymous;
-
-    /**
-     * @var bool
-     */
-    protected $isAnonymousVpn;
-
-    /**
-     * @var bool
-     */
-    protected $isHostingProvider;
-
-    /**
-     * @var bool
-     */
-    protected $isPublicProxy;
-
-    /**
-     * @var bool
-     */
-    protected $isResidentialProxy;
-
-    /**
-     * @var bool
-     */
-    protected $isTorExitNode;
-
-    /**
-     * @var string
-     */
-    protected $ipAddress;
-
-    /**
-     * @var string
-     */
-    protected $network;
+    protected bool $isAnonymous;
+    protected bool $isAnonymousVpn;
+    protected bool $isHostingProvider;
+    protected bool $isPublicProxy;
+    protected bool $isResidentialProxy;
+    protected bool $isTorExitNode;
+    protected string $ipAddress;
+    protected string $network;
 
     /**
      * @ignore