+++ src/Model/ConnectionType.php
@@ -20,20 +20,9 @@ use GeoIp2\Util;
  */
 class ConnectionType extends AbstractModel
 {
-    /**
-     * @var string|null
-     */
-    protected $connectionType;
-
-    /**
-     * @var string
-     */
-    protected $ipAddress;
-
-    /**
-     * @var string
-     */
-    protected $network;
+    protected ?string $connectionType;
+    protected string $ipAddress;
+    protected string $network;
 
     /**
      * @ignore