+++ src/Model/Isp.php
@@ -32,45 +32,14 @@ use GeoIp2\Util;
  */
 class Isp extends AbstractModel
 {
-    /**
-     * @var int|null
-     */
-    protected $autonomousSystemNumber;
-
-    /**
-     * @var string|null
-     */
-    protected $autonomousSystemOrganization;
-
-    /**
-     * @var string|null
-     */
-    protected $isp;
-
-    /**
-     * @var string|null
-     */
-    protected $mobileCountryCode;
-
-    /**
-     * @var string|null
-     */
-    protected $mobileNetworkCode;
-
-    /**
-     * @var string|null
-     */
-    protected $organization;
-
-    /**
-     * @var string
-     */
-    protected $ipAddress;
-
-    /**
-     * @var string
-     */
-    protected $network;
+    protected ?int $autonomousSystemNumber;
+    protected ?string $autonomousSystemOrganization;
+    protected ?string $isp;
+    protected ?string $mobileCountryCode;
+    protected ?string $mobileNetworkCode;
+    protected ?string $organization;
+    protected string $ipAddress;
+    protected string $network;
 
     /**
      * @ignore