+++ src/WebService/Client.php
@@ -51,17 +51,9 @@ class Client implements ProviderInterface
     /**
      * @var array<string>
      */
-    private $locales;
-
-    /**
-     * @var WsClient
-     */
-    private $client;
-
-    /**
-     * @var string
-     */
-    private static $basePath = '/geoip/v2.1';
+    private array $locales;
+    private WsClient $client;
+    private static string $basePath = '/geoip/v2.1';
 
     public const VERSION = 'v2.13.0';
 
@@ -135,6 +127,8 @@ class Client implements ProviderInterface
      * @throws \GeoIp2\Exception\GeoIp2Exception This serves as the parent
      *                                           class to the above exceptions. It will be thrown directly
      *                                           if a 200 status code is returned but the body is invalid.
+     * @throws \InvalidArgumentException         if something other than a single IP address or "me" is
+     *                                           passed to the method
      */
     public function city(string $ipAddress = 'me'): City
     {
@@ -165,6 +159,8 @@ class Client implements ProviderInterface
      * @throws \GeoIp2\Exception\GeoIp2Exception This serves as the parent class to the above exceptions. It
      *                                           will be thrown directly if a 200 status code is returned but
      *                                           the body is invalid.
+     * @throws \InvalidArgumentException         if something other than a single IP address or "me" is
+     *                                           passed to the method
      */
     public function country(string $ipAddress = 'me'): Country
     {
@@ -195,6 +191,8 @@ class Client implements ProviderInterface
      * @throws \GeoIp2\Exception\GeoIp2Exception This serves as the parent
      *                                           class to the above exceptions. It will be thrown directly
      *                                           if a 200 status code is returned but the body is invalid.
+     * @throws \InvalidArgumentException         if something other than a single IP address or "me" is
+     *                                           passed to the method
      */
     public function insights(string $ipAddress = 'me'): Insights
     {
@@ -204,6 +202,11 @@ class Client implements ProviderInterface
 
     private function responseFor(string $endpoint, string $class, string $ipAddress): Country
     {
+        if ($ipAddress !== 'me' && !filter_var($ipAddress, \FILTER_VALIDATE_IP)) {
+            throw new \InvalidArgumentException(
+                "The value \"$ipAddress\" is not a valid IP address."
+            );
+        }
         $path = implode('/', [self::$basePath, $endpoint, $ipAddress]);
 
         try {